atualização do gerenciamento de estoque para produtos em kg

+++ src/main/java/com/cefetmg/boinabrasa/dto/ProdutoRequestDTO.java
@@ -4,7 +4,6 @@ import lombok.Data;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-
 import java.math.BigDecimal;
 
 @Data
@@ -20,10 +19,10 @@ public class ProdutoRequestDTO {
     @NotBlank(message = "A unidade de referência do produto é obrigatória.")
     private String unidade;
 
-    private Integer quantidadeEstoque;
+    private BigDecimal quantidadeEstoque;
 
     @NotNull(message = "Verificação obrigatória")
     private Boolean controleEstoque;
 
-    private Boolean ativo; // isso aq e novo
+    private Boolean ativo; 
 }
